use req.originalUrl instead of req.url
because req.url can be overwritten, e.g. in case of hierarchical
sub-routers in node express. When using sub-routers (delegating routes to
child routers mounted at a specific mount point in the express app), the
sub-router will only receive the "sub-URL" which is the req.originalUrl,
with the mount point of the sub-router stripped off. So, for
cas-authentication redirects to work properly, req.originalUrl has to be
used throughout.